requirements.txt
tensorflow==2.12.0
Pillow==9.5.0
scikit-learn==1.2.2
opencv-python==4.7.0.72
matplotlib==3.7.1